fix/restructure-mobile-asset
